requirements.txt
flask>=3.0.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
cryptography>=42.0.0
argon2-cffi>=23.1.0
flask-limiter>=3.5.0
flask-wtf>=1.2.1
requests>=2.31.0
pytest>=8.0.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
